+++ src/listeners/panel/TaxonomyListener.ts
@@ -1,11 +1,20 @@
 import { CommandToCode } from '../../panelWebView/CommandToCode';
 import { TagType } from '../../panelWebView/TagType';
 import { BaseListener } from './BaseListener';
-import { window } from 'vscode';
-import { ArticleHelper, Settings } from '../../helpers';
+import { authentication, window } from 'vscode';
+import { ArticleHelper, Extension, Settings } from '../../helpers';
 import { BlockFieldData, CustomTaxonomyData, PostMessageData, TaxonomyType } from '../../models';
 import { DataListener } from '.';
-import { SETTING_TAXONOMY_CATEGORIES, SETTING_TAXONOMY_TAGS } from '../../constants';
+import {
+  DefaultFields,
+  SETTING_SEO_DESCRIPTION_FIELD,
+  SETTING_SEO_TITLE_FIELD,
+  SETTING_TAXONOMY_CATEGORIES,
+  SETTING_TAXONOMY_TAGS
+} from '../../constants';
+import { SponsorAi } from '../../services/SponsorAI';
+import { ExplorerView } from '../../explorerView/ExplorerView';
+import { MessageHandlerData } from '@estruyf/vscode';
 
 export class TaxonomyListener extends BaseListener {
   /**
@@ -52,7 +61,69 @@ export class TaxonomyListener extends BaseListener {
       case CommandToCode.addToCustomTaxonomy:
         this.addCustomTaxonomy(msg.payload);
         break;
+      case CommandToCode.aiSuggestTaxonomy:
+        this.aiSuggestTaxonomy(msg.command, msg.requestId, msg.payload);
+        break;
     }
+  }
+
+  private static async aiSuggestTaxonomy(command: string, requestId?: string, type?: TagType) {
+    if (!command || !requestId || !type) {
+      return;
+    }
+
+    const extPath = Extension.getInstance().extensionPath;
+    const panel = ExplorerView.getInstance(extPath);
+
+    const editor = window.activeTextEditor;
+    if (!editor) {
+      panel.getWebview()?.postMessage({
+        command,
+        requestId,
+        error: 'No active editor'
+      } as MessageHandlerData<string>);
+      return;
+    }
+
+    const article = ArticleHelper.getFrontMatter(editor);
+    if (!article || !article.data) {
+      panel.getWebview()?.postMessage({
+        command,
+        requestId,
+        error: 'No article data'
+      } as MessageHandlerData<string>);
+      return;
+    }
+
+    const githubAuth = await authentication.getSession('github', ['read:user'], { silent: true });
+    if (!githubAuth || !githubAuth.accessToken) {
+      return;
+    }
+
+    const titleField = (Settings.get(SETTING_SEO_TITLE_FIELD) as string) || DefaultFields.Title;
+    const descriptionField =
+      (Settings.get(SETTING_SEO_DESCRIPTION_FIELD) as string) || DefaultFields.Description;
+
+    const suggestions = await SponsorAi.getTaxonomySuggestions(
+      githubAuth.accessToken,
+      article.data[titleField] || '',
+      article.data[descriptionField] || '',
+      type
+    );
+
+    if (!suggestions) {
+      panel.getWebview()?.postMessage({
+        command,
+        requestId,
+        error: 'No article data'
+      } as MessageHandlerData<string>);
+    }
+
+    panel.getWebview()?.postMessage({
+      command,
+      requestId,
+      payload: suggestions || []
+    } as MessageHandlerData<string[]>);
   }
 
   /**