+++ src/panelWebView/components/Fields/FieldTitle.tsx
@@ -8,13 +8,15 @@ export interface IFieldTitleProps {
   icon?: JSX.Element;
   className?: string;
   required?: boolean;
+  actionElement?: JSX.Element;
 }
 
 export const FieldTitle: React.FunctionComponent<IFieldTitleProps> = ({
   label,
   icon,
   className,
-  required
+  required,
+  actionElement,
 }: React.PropsWithChildren<IFieldTitleProps>) => {
   const Icon = useMemo(() => {
     return icon ? React.cloneElement(icon, { style: { width: '16px', height: '16px' } }) : null;
@@ -23,9 +25,13 @@ export const FieldTitle: React.FunctionComponent<IFieldTitleProps> = ({
   return (
     <VsLabel>
       <div className={`metadata_field__label ${className || ''}`}>
-        {Icon}
-        <span style={{ lineHeight: '16px' }}>{label}</span>
-        <RequiredAsterix required={required} />
+        <div>
+          {Icon}
+          <span style={{ lineHeight: '16px' }}>{label}</span>
+          <RequiredAsterix required={required} />
+        </div>
+
+        {actionElement}
       </div>
     </VsLabel>
   );